// web/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        OrganisationDetail: {
            id: string;
            nameEl: string;
            nameEn: null | string;
            type: null | string;
            taxId: null | string;
            parent: null | components["schemas"]["OrganisationRef"];
            /** Format: uri */
            website: null | string;
            itemCounts: {
                [key: string]: number;
            };
            /** Format: double */
            awardedLast12MonthsEur: null | number;
            /** Format: double */
            paidLast12MonthsEur: null | number;
            awardedByYear: components["schemas"]["YearSpend"][];
            topContractors: components["schemas"]["ContractorSpend"][];
            recentItems: components["schemas"]["ItemSummary"][];
            dominantSupplier?: null | components["schemas"]["DominantSupplier"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}